requirements.txt
librosa
sounddevice
soundfile
pyttsx3
Pillow
matplotlib
pyswisseph
pandas
pytz
openpyxl
gunicorn
whitenoise
dj-database-url
psycopg2-binary
numpy
Django>=4.2
